% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagged_df.R
\name{create_lagged_df}
\alias{create_lagged_df}
\title{Create datasets with lagged features}
\usage{
create_lagged_df(data, type = c("train", "forecast"), outcome_cols = 1,
  horizons = NULL, lookback = NULL, lookback_control = NULL)
}
\arguments{
\item{data}{A data.frame with the target to be forecasted plus features/predictors.}

\item{type}{The type of dataset to return.}

\item{outcome_cols}{The column index of the target to be forecasted; only univariate forecasting is supported at present.}

\item{horizons}{Required. A vector of one or more forecast horizons, h, measured in dataset rows. For each horizon, 1:h
forecasts are returned (e.g., 'horizon = 12' trains a model to minimize 12-step-ahead error and returns forecasts
for 1:12 steps into the future).}

\item{lookback}{The lag--in dataset rows--for each predictor. Lags that don't support direct forecasting for a given
horizon are dropped. Either 'lookback' or 'lookback_control' need to be specified.}

\item{lookback_control}{A list of vectors, specifying potentially unique lags for each predictor. The length
of the list should equal ncol(data) and be ordered the same as the columns in data). Lags that don't support
direct forecasting to 'horizon' will throw an error.}
}
\value{
A 'lagged_df' object: A list of data.frames with new columns for the lagged predictors.
}
\description{
Create a list of datasets with lagged predictors to (a) train forecasting models for
specified forecast horizons and (b) forecast with lagged predictors.
}
\examples{
\donttest{
# Sampled Seatbelts data from the R package datasets.
data("data_seatbelts", package = "forecastML")
#------------------------------------------------------------------------------
# Example 1 - Training data for 12 horizon-specific models w/ common lags per predictor.
horizons <- 1:12
lookback <- 1:15

data_train <- create_lagged_df(data_seatbelts, type = "train", outcome_cols = 1,
                               lookback = lookback, horizon = horizons)
head(data_train[[length(horizons)]])

# Example 1 - Forecasting dataset
# The last 'nrow(data_seatbelts) - horizon' rows are automatically used from data_seatbelts.
data_forecast <- create_lagged_df(data_seatbelts, type = "forecast", outcome_cols = 1,
                                  lookback = lookback, horizon = horizons)
head(data_forecast[[length(horizons)]])
#------------------------------------------------------------------------------
# Example 2 - Training data for one 3-month horizon model w/ unique lags per predictor.
horizons <- 3
lookback <- list(c(3, 6, 9, 12), c(4:12), c(6:15), c(8))

data_train <- create_lagged_df(data_seatbelts, type = "train", outcome_cols = 1,
                               lookback_control = lookback, horizon = horizons)
head(data_train[[length(horizons)]])
#------------------------------------------------------------------------------
}
}
