% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_seatbelts.R
\docType{data}
\name{data_seatbelts}
\alias{data_seatbelts}
\title{Road Casualties in Great Britain 1969-84}
\format{
A data.frame with 192 rows and 8 columns
}
\source{
Harvey, A.C. (1989). Forecasting, Structural Time Series Models and the Kalman Filter. Cambridge University Press, pp. 519–523.

Durbin, J. and Koopman, S. J. (2001). Time Series Analysis by State Space Methods. Oxford University Press.

\url{https://stat.ethz.ch/R-manual/R-devel/library/datasets/html/UKDriverDeaths.html}
}
\usage{
data_seatbelts
}
\description{
This is the \code{Seatbelts} dataset from the \code{datasets} package.
}
\keyword{datasets}
