% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_gaps.R
\name{fill_gaps}
\alias{fill_gaps}
\title{Prepare a dataset for modeling by filling in temporal gaps in data collection}
\usage{
fill_gaps(data, date_col = 1, frequency = NULL, groups = NULL,
  static_features = NULL)
}
\arguments{
\item{data}{A data.frame with, minimally, dates and the outcome being forecasted.}

\item{date_col}{The column index--an integer--of the date index. This column should have class 'Date'.}

\item{frequency}{A string taking the same input as \code{base::seq(..., by = "frequency")} e.g., '1 month', '7 days', etc.}

\item{groups}{Column name(s) that identify the groups/hierarchies when multiple time-series are present. These columns are used as model predictors but are not lagged.
Note that combining feature lags with grouped time-series will result in NA values throughout the data.}

\item{static_features}{For grouped time-series only. Column name(s) that identify features that do not change through time.
These columns are used as model features but are not lagged.
Note that combining feature lags with grouped time-series will result in NA values throughout the data.}
}
\value{
An object of class 'data.frame': The returned data.frame has the same number of columns and column order but
with additional rows to account for gaps in data collection. If the user-supplied
forecasting algorithm(s) cannot handle missing outcome values or missing dynamic features, these should either be
imputed prior to \code{create_lagged_df} or filtered out in the user-supplied modeling function for \code{\link{train_model}}.
}
\description{
In order to create a modeling dataset with feature lags that are temporally correct, the entry
function in \code{forecastML}, \code{\link{create_lagged_df}}, needs evenly-spaced time-series with no
gaps in data collection. \code{fill_gaps()} can help here.
This function takes a \code{data.frame} with (a) dates, (b) the outcome being forecasted, and, optionally,
(c) dynamic features that change through time, (d) group columns for multiple time-series modeling,
and (d) static or non-dynamic features for multiple time-series modeling and returns a \code{data.frame}
with rows evenly spaced in time. Specifically, this function adds rows to the input dataset
while filling in (a) dates, (b) grouping information, and (c) static features. The (a) outcome and (b)
dynamic features will be \code{NA} for any missing time periods; these \code{NA} values can be left
as-is, user-imputed, or removed from modeling in the user-supplied modeling wrapper function for \code{\link{train_model}}.
}
\section{Methods and related functions}{


The output of \code{fill_gaps()} is passed into

\itemize{
  \item \code{\link{create_lagged_df}}
}
}

\examples{
\donttest{
# Sampled Seatbelts data from the R package datasets.
data("data_seatbelts", package = "forecastML")
#------------------------------------------------------------------------------
# Example 1 - Training data for 12 horizon-specific models w/ common lags per predictor.
horizons <- 1:12
lookback <- 1:15

data <- data_seatbelts

data_train <- create_lagged_df(data_seatbelts, type = "train", outcome_cols = 1,
                               lookback = lookback, horizon = horizons)
head(data_train[[length(horizons)]])

# Example 1 - Forecasting dataset
# The last 'nrow(data_seatbelts) - horizon' rows are automatically used from data_seatbelts.
data_forecast <- create_lagged_df(data_seatbelts, type = "forecast", outcome_cols = 1,
                                  lookback = lookback, horizon = horizons)
head(data_forecast[[length(horizons)]])
#------------------------------------------------------------------------------
# Example 2 - Training data for one 3-month horizon model w/ unique lags per predictor.
horizons <- 3
lookback <- list(c(3, 6, 9, 12), c(4:12), c(6:15), c(8))

data_train <- create_lagged_df(data_seatbelts, type = "train", outcome_cols = 1,
                               lookback_control = lookback, horizon = horizons)
head(data_train[[length(horizons)]])
#------------------------------------------------------------------------------
}
}
