% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_error.R
\name{plot.forecast_error}
\alias{plot.forecast_error}
\title{Plot forecast error}
\usage{
\method{plot}{forecast_error}(
  x,
  data_results,
  type = c("window", "horizon", "global"),
  metric = NULL,
  facet = NULL,
  models = NULL,
  horizons = NULL,
  windows = NULL,
  group_filter = NULL,
  ...
)
}
\arguments{
\item{x}{An object of class 'validation_error' from \code{return_error()}.}

\item{data_results}{An object of class 'forecast_results' from \code{predict.forecast_model()}.}

\item{type}{Select plot type; \code{type = "window"} is the default plot.}

\item{metric}{Select error metric to plot (e.g., "mae"); \code{attributes(x)$error_metrics[1]} is the default metric.}

\item{facet}{Optional. A formula with any combination of \code{horizon}, \code{model}, or \code{group} (for grouped time series).
passed to \code{ggplot2::facet_grid()} internally (e.g., \code{horizon ~ model}, \code{horizon + model ~ .}, \code{~ horizon + group}).
Can be \code{NULL}. The default faceting is set internally depending on the plot \code{type}.}

\item{models}{Optional. A vector of user-defined model names from \code{train_model()} to filter results.}

\item{horizons}{Optional. A numeric vector to filter results by horizon.}

\item{windows}{Optional. A numeric vector to filter results by validation window number.}

\item{group_filter}{A string for filtering plot results for grouped time-series (e.g., \code{"group_col_1 == 'A'"}).}

\item{...}{Not used.}
}
\value{
Forecast error plots of class 'ggplot'.
}
\description{
Plot forecast error at various levels of aggregation.
}
