% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_model.R
\name{plot.forecast_results}
\alias{plot.forecast_results}
\title{Plot an object of class forecast_results}
\usage{
\method{plot}{forecast_results}(forecast_results, data_actual = NULL,
  models = NULL, horizons = NULL, windows = NULL,
  facet_plot = c("model", "model_forecast_horizon"))
}
\arguments{
\item{forecast_results}{An object of class 'forecast_results' from predict.forecast_model().}

\item{data_actual}{A data.frame containing the target/outcome name; other columns are ignored.
The data can be historical and/or holdout/test data, forecasts and actuals are matched by row.names().}

\item{models}{Filter results by user-defined model name from train_model() (optional).}

\item{horizons}{Filter results by horizon (optional).}

\item{windows}{Filter results by validation window number (optional).}

\item{facet_plot}{Adjust the plot display through ggplot2::facet_grid(). facet_plot = NULL plots results in one facet.}
}
\value{
Forecast plot of class 'ggplot'.
}
\description{
A forecast plot for each horizon for each model in predict.forecast_model().
}
