% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_model.R
\name{plot.training_results}
\alias{plot.training_results}
\title{Plot an object of class training_results}
\usage{
\method{plot}{training_results}(x, type = c("prediction", "residual",
  "forecast_stability", "forecast_variability"), models = NULL,
  horizons = NULL, windows = NULL, valid_indices = NULL,
  group_filter = NULL, ...)
}
\arguments{
\item{x}{An object of class 'training_results' from \code{predict.forecast_model}.}

\item{type}{Plot type, default is "prediction" for hold-out sample predictions.}

\item{models}{Filter results by user-defined model name from \code{train_model} (optional).}

\item{horizons}{Filter results by horizon (optional).}

\item{windows}{Filter results by validation window number (optional).}

\item{valid_indices}{Filter results by validation row indices or dates (optional).}

\item{group_filter}{A string for filtering plot results for grouped time-series (e.g., "group_col_1 == 'A'").}

\item{...}{Arguments passed to \code{base::plot}}
}
\value{
Diagnostic plots of class 'ggplot'.
}
\description{
Several diagnostic plots can be returned to assess the quality of the forecats
based on predictions on the outer-loop validation datasets.
}
