% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_error.R
\name{plot.validation_error}
\alias{plot.validation_error}
\title{Plot validation dataset forecast error}
\usage{
\method{plot}{validation_error}(data_error, data_results,
  type = c("time", "horizon", "global"), models = NULL,
  horizons = NULL, window_lengths = NULL, windows = NULL)
}
\arguments{
\item{data_error}{An object of class 'validation_error' from return_error().}

\item{data_results}{An object of class 'training_results' from predict.forecast_model().}

\item{type}{Select plot type; 'time' is the default plot.}

\item{models}{Filter results by user-defined model name from train_model() (optional).}

\item{horizons}{Filter results by horizon (optional).}

\item{windows}{Filter results by validation window number (optional).}
}
\value{
Forecast error plots of class 'ggplot'.
}
\description{
Plot forecast error at various levels of aggregation across validation datasets.
}
