% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_windows.R
\name{plot.windows}
\alias{plot.windows}
\title{Plot validation datasets}
\usage{
\method{plot}{windows}(windows, data, show_labels = TRUE)
}
\arguments{
\item{windows}{An object of class 'windows' from create_windows().}

\item{data}{An object of class 'lagged_df' from create_lagged_df().}

\item{show_labels}{Show validation dataset IDs on the plot.}
}
\value{
A plot of the outer-loop nested cross-validation windows.
}
\description{
Plot validation datasets across time.
}
\examples{
\donttest{
# Sampled Seatbelts data from the R package datasets.
data("data_seatbelts", package = "forecastML")

# Example - Training data for 12 horizon-specific models w/ common lags per predictor.
horizons <- 1:12
lookback <- 1:15

data_train <- create_lagged_df(data_seatbelts, type = "train", outcome_cols = 1,
                               lookback = lookback, horizon = horizons)

windows <- create_windows(data_train, window_length = 12)

plot(windows, data_train)
}
}
