% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lagged_df.R
\name{summary.lagged_df}
\alias{summary.lagged_df}
\title{Return a summary of a lagged_df object}
\usage{
\method{summary}{lagged_df}(object, ...)
}
\arguments{
\item{object}{An object of class 'lagged_df' from \code{create_lagged_df}.}

\item{...}{Arguments passed to \code{base::summary}.}
}
\value{
A printed summary of the contents of the lagged_df object.
}
\description{
Return a summary of a lagged_df object
}
