% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_model.R
\name{ts_reg}
\alias{ts_reg}
\title{Forecast with Regression Model}
\usage{
ts_reg(input, y = NULL, x = NULL, seasonal = NULL,
  trend = list(linear = TRUE, exponential = FALSE, log = FALSE, power =
  FALSE), lags = NULL, method = "lm", method_arg = list(step = FALSE,
  direction = "both"), scale = NULL)
}
\arguments{
\item{input}{A tsibble or ts object}

\item{y}{A character, the column name of the depended variable of the input object, required (and applicable) only when the input is tsibble object}

\item{x}{A character, the column names of the independed variable of the input object, applicable when using tsibble object with regressors}

\item{seasonal}{A character, optional, create categorical variable/s to model the single or multiple seasonal components. Supporting the
following frequencies structure c("quarter", "month", "week", "yday", "wday", "hour") for quarterly,
monthly, weekly, day of the year, day of the week, and hourly respectively}

\item{trend}{A list, define the trend structure. Possible arguments -
"linear", a boolean variable, if set to TRUE defines a linear trend (e.g., index of 1,2,3,...,t, for a series with t observations)
"power", an numeric value, defines the polynomial degree of the series index (for example a power = 0.5 define a square root index and power = 2 defines a squared index).
"exponential" - a boolean variable, if set to TRUE defines an exponential trend.
"log" - a boolean variable, if set to TRUE defines a log transformation for the trend.
By default, the trend argument is set to a linear trend (i.e., power = 1)}

\item{lags}{A positive integer, defines the series lags to be used as input to the model (equivalent to AR process)}

\item{method}{A character, defines the regression method to be used, currently only "lm" method is available}

\item{method_arg}{A list, defines the argument of the selected method}

\item{scale}{A character, scaling options of the series, methods available -
c("log", "normal", "standard") for log transformation, normalization, or standardization of the series, respectively.
If set to NULL (default), no transformation will occur}
}
\description{
Forecasting regular time series data with regression models
}
