% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util function.R
\name{tsACF}
\alias{tsACF}
\title{Autocorrelation Function}
\usage{
tsACF(input, var = NULL, max.lag = NULL, ci = 0.95, na.rm = FALSE)
}
\arguments{
\item{input}{A tsibble object}

\item{var}{A character, optional, defines the variables names to calculate the ACF when having a multuple time series object}

\item{max.lag}{An integer, defines the maximum number of lags to be used}
}
\description{
Autocorrelation Function
}
