% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_windows.R
\name{create_windows}
\alias{create_windows}
\title{Create validation datasets for the outer loop of nested cross-validation}
\usage{
create_windows(lagged_df, window_length = 12, window_start = NULL,
  window_stop = NULL, skip = 0, include_partial_window = TRUE)
}
\arguments{
\item{lagged_df}{An object of class 'lagged_df' from create_lagged_df().}

\item{window_length}{An integer that defines the length of the contiguous validation dataset in dataset rows.
Setting 'window_length = 0' trains the model on the entire dataset--useful for re-training after examining
the nested cross-validation results.}

\item{window_start}{An optional index identifying the row to start creating contiguous validation datasets.}

\item{window_stop}{An optional index identifying the row to stop creating contiguous validation datasets.}

\item{skip}{An integer giving a fixed number of dataset rows to skip between validation datasets.}

\item{include_partial_window}{Keep validation datasets that are shorter than window_length.}
}
\value{
A'windows' object: A list of matrices, one per horizon, giving the indices for the validation datasets.
}
\description{
Create blocks of contiguous validation windows to assess the likely forecast accuracy
of various models.
}
\examples{
\donttest{
# Sampled Seatbelts data from the R package datasets.
data("data_seatbelts", package = "forecastML")

# Example - Training data for 12 horizon-specific models w/ common lags per predictor.
horizons <- 1:12
lookback <- 1:15

data_train <- create_lagged_df(data_seatbelts, type = "train", outcome_cols = 1,
                               lookback = lookback, horizon = horizons)

create_windows(data_train, window_length = 12)
}
}
