% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_gaps.R
\name{fill_gaps}
\alias{fill_gaps}
\title{Prepare a dataset for modeling by filling in temporal gaps in data collection}
\usage{
fill_gaps(data, date_col = 1, frequency = NULL, groups = NULL,
  static_features = NULL)
}
\arguments{
\item{data}{A data.frame with, minimally, dates and the outcome being forecasted.}

\item{date_col}{The column index--an integer--of the date index. This column should have class 'Date'.}

\item{frequency}{A string taking the same input as \code{base::seq(..., by = "frequency")} e.g., '1 month', '7 days', etc.}

\item{groups}{Column name(s) that identify the groups/hierarchies when multiple time-series are present. These columns are used as model predictors but are not lagged.
Note that combining feature lags with grouped time-series will result in NA values throughout the data.}

\item{static_features}{For grouped time-series only. Column name(s) that identify features that do not change through time.
These columns are used as model features but are not lagged.
Note that combining feature lags with grouped time-series will result in NA values throughout the data.}
}
\value{
An object of class 'data.frame': The returned data.frame has the same number of columns and column order but
with additional rows to account for gaps in data collection. If the user-supplied
forecasting algorithm(s) cannot handle missing outcome values or missing dynamic features, these should either be
imputed prior to \code{create_lagged_df} or filtered out in the user-supplied modeling function for \code{\link{train_model}}.
}
\description{
In order to create a modeling dataset with feature lags that are temporally correct, the entry
function in \code{forecastML}, \code{\link{create_lagged_df}}, needs evenly-spaced time-series with no
gaps in data collection. \code{fill_gaps()} can help here.
This function takes a \code{data.frame} with (a) dates, (b) the outcome being forecasted, and, optionally,
(c) dynamic features that change through time, (d) group columns for multiple time-series modeling,
and (d) static or non-dynamic features for multiple time-series modeling and returns a \code{data.frame}
with rows evenly spaced in time. Specifically, this function adds rows to the input dataset
while filling in (a) dates, (b) grouping information, and (c) static features. The (a) outcome and (b)
dynamic features will be \code{NA} for any missing time periods; these \code{NA} values can be left
as-is, user-imputed, or removed from modeling in the user-supplied modeling wrapper function for \code{\link{train_model}}.
}
\section{Methods and related functions}{


The output of \code{fill_gaps()} is passed into

\itemize{
  \item \code{\link{create_lagged_df}}
}
}

\examples{
\donttest{

  # NOAA buoy dataset with gaps in data collection
  data("data_buoy_gaps", package = "forecastML")

  data_buoy_no_gaps <- fill_gaps(data_buoy_gaps, date_col = 1, frequency = '1 day',
                                 groups = 'buoy_id', static_features = c('lat', 'lon'))

  # The returned data.frame has the same number of columns but the time-series
  # are now evenly spaced at 1 day apart. Additionally, the unchanging grouping
  # columns and static features columns have been filled in for the newly created dataset rows.
  dim(data_buoy_gaps)
  dim(data_buoy_no_gaps)

  # Running create_lagged_df() is the next step in the forecastML forecasting
  # process. If there are long gaps in data collection, like in this buoy dataset,
  # and the user-supplied modeling algorithm cannot handle missing outcomes data,
  # the best option is to filter these rows out in the user-supplied modeling function
  # for train_model()
}
}
