% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_model.R
\name{forecastML}
\alias{forecastML}
\title{Forecast trainML Model}
\usage{
forecastML(model, newdata = NULL, h, pi = c(0.95, 0.8))
}
\arguments{
\item{model}{A trainML object}

\item{newdata}{A tsibble object, must be used when the input model was trained with external inputs (i.e., the 'x' argument of the trainML function was used). This input must follow the following structure:

- Use the same time intervals (monthly, daily, hourly, etc.) structure and timestamp class (e.g., yearquarter, yearmonth, POSIXct, etc.) as the original input

- The number of observations must align with the forecasting horizon (the 'h' argument)

-  The timestamp of the first observation must be the consecutive observation of the last observation of the original series}

\item{h}{An integer, define the forecast horizon}

\item{pi}{A vector with numeric values between 0 and 1, define the level of the confidence of the prediction intervals of the forecast. By default calculate the 80% and 95% prediction intervals}
}
\description{
Forecast trainML Model
}
