% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_model.R
\name{plot.training_results}
\alias{plot.training_results}
\title{Plot an object of class training_results}
\usage{
\method{plot}{training_results}(x, type = c("prediction", "residual",
  "forecast_stability", "forecast_variability"), models = NULL,
  horizons = NULL, windows = NULL, valid_indices = NULL,
  group_filter = NULL, ...)
}
\arguments{
\item{x}{An object of class 'training_results' from \code{predict.forecast_mode()l}.}

\item{type}{Plot type, default is "prediction", for hold-out sample predictions.}

\item{models}{Optional. Filter results by user-defined model name from \code{train_model()}.}

\item{horizons}{Optional. A numeric vector of model forecast horizons to filter results by horizon-specific model results.}

\item{windows}{Optional. A numeric vector of windows to filter results by validation window number.}

\item{valid_indices}{Optional. A numeric or date vector to filter results by validation row indices or dates.}

\item{group_filter}{Optional. A string for filtering plot results for grouped time-series
(e.g., \code{"group_col_1 == 'A'"}). The results are passed to \code{dplyr::filter()} internally.}

\item{...}{Arguments passed to \code{base::plot()}}
}
\value{
Diagnostic plots of class 'ggplot'.
}
\description{
Several diagnostic plots can be returned to assess the quality of the forecasts
based on predictions on the validation datasets.
}
