% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_error.R
\name{plot.validation_error}
\alias{plot.validation_error}
\title{Plot validation dataset forecast error}
\usage{
\method{plot}{validation_error}(x, data_results, type = c("time",
  "horizon", "global"), models = NULL, horizons = NULL,
  windows = NULL, group_filter = NULL, ...)
}
\arguments{
\item{x}{An object of class 'validation_error' from \code{return_error()}.}

\item{data_results}{An object of class 'training_results' from \code{predict.forecast_model()}.}

\item{type}{Select plot type; \code{type = "time"} is the default plot.}

\item{models}{Optional. A vector of user-defined model names from \code{train_model()} to filter results.}

\item{horizons}{Optional. A numeric vector to filter results by horizon.}

\item{windows}{Optional. A numeric vector to filter results by validation window number.}

\item{group_filter}{A string for filtering plot results for grouped time-series (e.g., \code{"group_col_1 == 'A'"}).}

\item{...}{Not used.}
}
\value{
Forecast error plots of class 'ggplot'.
}
\description{
Plot forecast error at various levels of aggregation across validation datasets.
}
