% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_fc}
\alias{plot_fc}
\title{Plotting the forecast output}
\usage{
plot_fc(forecast, theme = "normal")
}
\arguments{
\item{forecast}{A forecastML object}

\item{theme}{A character, defines the color theme to be used in the plot output.
Available themes - "normal" (default), "darkBlue", "darkPink", "darkGreen", "classic", "lightBeige"}
}
\value{
A plotly object
}
\description{
Plotting the forecast output
}
