% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_res}
\alias{plot_res}
\title{Plotting the residuals}
\usage{
plot_res(model, na.rm = FALSE, margin = 0.04)
}
\arguments{
\item{model}{A trainML object}

\item{na.rm}{A boolean, if TRUE will ignore missing values, by default set to FALSE}

\item{margin}{A numeric, define the margin space between the subplot components}
}
\description{
Plotting the residuals
}
\examples{

# Train a time series forecasting model
md <- trainML(input = AirPassengers, trend = list(linear = TRUE), seasonal = "month")

# create a residuals plot
plot_res(md)
}
