% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_model.R
\name{residuals}
\alias{residuals}
\title{Return model residuals}
\usage{
residuals(x, ...)
}
\arguments{
\item{x}{An object of class 'training_results' from running \code{predict()} on a training dataset.}

\item{...}{Not used.}
}
\value{
A data.frame of model residuals of class 'training_residuals'.
}
\description{
Return model residuals
}
