% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_model.R
\name{ts_reg}
\alias{ts_reg}
\title{Forecast with Regression Model}
\usage{
ts_reg(input, y = NULL, x = NULL, seasonal = NULL,
  trend = list(power = c(1), exponential = FALSE, log = FALSE),
  lags = NULL, method = "lm", method_arg = list(step = FALSE,
  direction = "both"), scale = NULL)
}
\arguments{
\item{input}{A tsibble or ts object}

\item{y}{A character, the column name of the depended variable of the input object}
}
\description{
Visualization of series with its lags,
can be used to identify a correlation between the series and it lags
}
